refactor: componentsの読み込み設定
