+++ src/resources/js/app.js
@@ -8,13 +8,6 @@ components.forEach(component => {
   Vue.component(component.name, component.object)
 })
 
-// form
-Vue.component('form-wrapper', require('./moleclues/form/form-wrapper.vue').default);
-Vue.component('input-form', require('./moleclues/form/input-form.vue').default);
-Vue.component('range-input-form', require('./moleclues/form/range-input-form.vue').default);
-Vue.component('select-form', require('./moleclues/form/select-form.vue').default);
-Vue.component('text-area-form', require('./moleclues/form/text-area-form.vue').default);
-
 import vuetify from './plugins/vuetify'
 
 const app = new Vue({
