+++ src/resources/js/molecules/form/range-input-form.vue
