+++ src/resources/js/molecules/landing-page/footer/social-constants.js
