+++ src/resources/js/molecules/landing-page/footer/social-icons.vue
