+++ src/resources/js/molecules/landing-page/mian-view.vue
