+++ src/resources/js/plugins/components/molecules/form-components.js
@@ -1,0 +1,22 @@
+export const formComponents = [
+  {
+    name: 'form-wrapper',
+    object: () => import('../../../molecules/form/form-wrapper.vue')
+  },
+  {
+    name: 'input-form',
+    object: () => import('../../../molecules/form/input-form.vue')
+  },
+  {
+    name: 'range-input-form',
+    object: () => import('../../../molecules/form/range-input-form.vue')
+  },
+  {
+    name: 'select-form',
+    object: () => import('../../../molecules/form/select-form.vue')
+  },
+  {
+    name: 'text-area-form',
+    object: () => import('../../../molecules/form/text-area-form.vue')
+  }
+]
